requirements.txt
https://github.com/Rapptz/discord.py/archive/rewrite.zip
PyYAML
